pyproject.toml
[project]
name = "chat-aggregator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.118.3",
    "fastmcp>=2.13.3",
    "google-adk>=1.20.0",
    "google-api-python-client>=2.187.0",
    "google-auth-httplib2>=0.2.1",
    "google-auth-oauthlib>=1.2.3",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "pylint>=4.0.4",
]
